refactor: clean up main by moving handlers out
