+++ cmd/web/handlers.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"fmt"
-	"log"
 	"net/http"
 	"strconv"
 )
@@ -12,13 +11,8 @@ func home(w http.ResponseWriter, r *http.Request) {
 		http.NotFound(w, r)
 		return
 	}
-	// for k, v := range w.Header() {
-	// 	fmt.Printf("key: %s      value: %s\n", k, v)
-	// }
-	w.Header()["Date"] = nil
-	w.Write([]byte("Hello from Anh Vo"))
+	w.Write([]byte("Hello from Snippetbox"))
 }
-
 func snippetView(w http.ResponseWriter, r *http.Request) {
 	id, err := strconv.Atoi(r.URL.Query().Get("id"))
 	if err != nil || id < 1 {
@@ -27,24 +21,11 @@ func snippetView(w http.ResponseWriter, r *http.Request) {
 	}
 	fmt.Fprintf(w, "Display a specific snippet with ID %d...", id)
 }
-
 func snippetCreate(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodPost {
 		w.Header().Set("Allow", http.MethodPost)
 		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
 		return
 	}
-
 	w.Write([]byte("Create a new snippet..."))
-}
-
-func main() {
-	mux := http.NewServeMux()
-	mux.HandleFunc("/", home)
-	mux.HandleFunc("/snippet/view", snippetView)
-	mux.HandleFunc("/snippet/create", snippetCreate)
-
-	log.Print("Starting server on :8080")
-	err := http.ListenAndServe(":8080", mux)
-	log.Fatal(err)
 }
